+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to cssdb
 
+### Unreleased (minor)
+
+- Added `interoperable_at` to make it possible to calculate the Baseline status of a feature
+- Updated `@mdn/browser-compat-data` to `5.5.2` (patch)
+- Updated `caniuse-lite` to `1.0.30001571` (patch)
+
 ### 7.9.1 (December 11, 2023)
 
 - Updated `@mdn/browser-compat-data` to `5.4.5` (patch)